+++ packages/amplify-category-hosting/lib/S3AndCloudFront/index.js
@@ -86,7 +86,6 @@ function removeCDN(context) {
 }
 
 function makeBucketPrivate(context) {
-  delete context.exeInfo.template.Resources.BucketPolicy;
   delete context.exeInfo.template.Resources.S3Bucket.Properties.AccessControl;
 }