+++ packages/amplify-category-auth/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-auth@1.0.5...amplify-category-auth@1.0.6) (2019-02-26)
+
+
+### Bug Fixes
+
+* **amplify-category-auth:** update auth cfn template to quote string ([1ff9e16](https://github.com/aws-amplify/amplify-cli/commit/1ff9e16)), closes [#882](https://github.com/aws-amplify/amplify-cli/issues/882)
+
+
+
+
+
 ## [1.0.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-auth@1.0.3-beta.0...amplify-category-auth@1.0.5) (2019-02-11)
 
 **Note:** Version bump only for package amplify-category-auth